Separated variable constants from schema getter constants
